dev-requirements.txt
pyOpenSSL
pytest
jsonschema
requests==2.31.0
requests-toolbelt==0.9
urllib3<2.0.0
myst-parser
PyJWT
jwcrypto
